requirements.txt
Django
djangorestframework
python-gitlab